pyproject.toml
[project]
name = "aim"
version = "0.0.1"
description = "An Information Management"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "yzx9", email = "pypi@yzx9.xyz" }]
keywords = []
license = { text = "Apache-2.0" }
dependencies = [
    "aiohttp[speedups]>=3.11.11",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "click>=8.1.8",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy[asyncio]>=2.0.37",
]

[project.urls]
repository = "https://github.com/yzx9/aim"

[dependency-groups]
dev = [
    "aiohttp-debugtoolbar>=0.6.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"
